Staging

+++ src/main/java/com/mysawit/pembayaran/dto/response/TopUpResponse.java
@@ -4,6 +4,7 @@ import com.mysawit.pembayaran.model.enums.TopUpStatus;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -13,8 +14,8 @@ public class TopUpResponse {
 
     private UUID id;
     private UUID userId;
-    private double amountRupiah;
-    private double amountSawitDollar;
+    private BigDecimal amountRupiah;
+    private BigDecimal amountSawitDollar;
     private String paymentGatewayRef;
     private String paymentUrl;
     private TopUpStatus status;
